Fixed Additional Info with additional HTML tags
The "Additional Info" section also contains the ```<b>``` and ```</b>```
tags, which were included in the ```METAL ARCHIVES INFO``` tag field.
This fix removes those tags.

Also, on the "Search By Band + Album" script it has been implemented a
fix to remove the whitespaces preceding the "]", and the "]" itself at
the end of the Album Type.